Construct the tuning space, i.e. myKernels accumulatively
